+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Classify some messages as warning.
 - Skip recently refreshed channels.
 - Group channels using tags.
+- Show warning if trying to subscribe to an already subscribed channel.
 
 ### Changed
 - Ignore case when sorting channels list